% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{data_example}
\alias{data_example}
\title{An example simulated dataset to use in examples}
\description{
An example simulated dataset to use in examples
}
\examples{
# Code used to generate data_example
set.seed(98123)
data_example = data.frame(
  Group = c(rep("Test", 30), rep("Reference", 30)),
  value = c(rnorm(n = 30, mean = 0, sd = .7),
            rnorm(n = 30, mean = 0, sd = 1)))
}
\keyword{data}
