% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.testTail}
\alias{plot.testTail}
\title{Generic plot function for testTail objects}
\usage{
\method{plot}{testTail}(x, ...)
}
\arguments{
\item{x}{An object resulting from \link{testis.tails}}

\item{...}{options to pass to the generic plot function.}
}
\description{
Generic plot function for testTail objects
}
\examples{
library(tailTest)
data(data_example)
head(data_example)
reference = data_example[which(data_example$Group == "Reference"), "value"]
test = data_example[which(data_example$Group == "Test"), "value"]
result    = testis.tails(Test = test, Reference = reference,
                         q = 0.1, alpha = 0.05, c = 20)
plot(result)

}
