% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.testTail}
\alias{print.testTail}
\title{Generic print function for testTail objects}
\usage{
\method{print}{testTail}(x, ...)
}
\arguments{
\item{x}{An object resulting from \link{testis.tails}}

\item{...}{options to pass to the generic print function.}
}
\description{
Generic print function for testTail objects
}
