% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{testis.tails}
\alias{testis.tails}
\title{given observations for Test and Reference,
the function calculates the test statistic with the tail-test}
\usage{
testis.tails(Test, Reference, q = 0.1, alpha = 0.05, c = NULL)
}
\arguments{
\item{Test}{vector of observations from Test}

\item{Reference}{vector of observations from Reference}

\item{q}{which quantil should be used (e.g., q = 0.05). Default to 0.10}

\item{alpha}{significance level. Default to 0.05}

\item{c}{equivalence margin. Default corresponds to the FDA's null hypothesis}
}
\description{
given observations for Test and Reference,
the function calculates the test statistic with the tail-test
}
\examples{
library(tailTest)
data(data_example)
head(data_example)
reference = data_example[which(data_example$Group == "Reference"), "value"]
test = data_example[which(data_example$Group == "Test"), "value"]
testis.tails(Test = test, Reference = reference,
             q = 0.1, alpha = 0.05, c = 20)
}
